Add inactive users notification and deletion jobs

+++ udata/core/user/models.py
@@ -237,7 +237,7 @@ class User(WithMetrics, UserMixin, db.Document):
         raise NotImplementedError("""This method should not be using directly.
         Use `mark_as_deleted` (or `_delete` if you know what you're doing)""")
 
-    def mark_as_deleted(self, notify: bool = True):
+    def mark_as_deleted(self, notify: bool = True, delete_comments: bool = True):
         if self.avatar.filename is not None:
             storage = storages.avatars
             storage.delete(self.avatar.filename)
@@ -265,16 +265,17 @@ class User(WithMetrics, UserMixin, db.Document):
                 member for member in organization.members if member.user != self
             ]
             organization.save()
-        for discussion in Discussion.objects(discussion__posted_by=self):
-            # Remove all discussions with current user as only participant
-            if all(message.posted_by == self for message in discussion.discussion):
-                discussion.delete()
-                continue
+        if delete_comments:
+            for discussion in Discussion.objects(discussion__posted_by=self):
+                # Remove all discussions with current user as only participant
+                if all(message.posted_by == self for message in discussion.discussion):
+                    discussion.delete()
+                    continue
 
-            for message in discussion.discussion:
-                if message.posted_by == self:
-                    message.content = "DELETED"
-            discussion.save()
+                for message in discussion.discussion:
+                    if message.posted_by == self:
+                        message.content = "DELETED"
+                discussion.save()
         Follow.objects(follower=self).delete()
         Follow.objects(following=self).delete()
